+++ src/editor/Warp.NET.Editor/Layouts/MainLayout.cs
@@ -1,3 +1,4 @@
+using Warp.NET.Numerics;
 using Warp.NET.RenderImpl.Ui.Components;
 using Warp.NET.RenderImpl.Ui.Components.Styles;
 using Warp.NET.Ui;
@@ -8,7 +9,8 @@ public class MainLayout : Layout
 {
 	public MainLayout()
 	{
-		Button button = new(new NormalizedBounds(0.1f, 0.1f, 0.1f, 0.1f), () => {}, ButtonStyle.Default);
+		ButtonStyle buttonStyle = new(Color.Black, Color.White, Color.Gray(0.5f), 2);
+		Button button = new(new NormalizedBounds(0.1f, 0.1f, 0.1f, 0.1f), () => { }, buttonStyle);
 		NestingContext.Add(button);
 	}
 }