+++ src/samples/Warp.NET.Samples.Ui/Layouts/MainLayout.cs
@@ -1,6 +1,8 @@
 using Warp.NET.Numerics;
 using Warp.NET.RenderImpl.Ui.Components;
 using Warp.NET.RenderImpl.Ui.Components.Styles;
+using Warp.NET.RenderImpl.Ui.Rendering.Text;
+using Warp.NET.Text;
 using Warp.NET.Ui;
 
 namespace Warp.NET.Samples.Ui.Layouts;
@@ -9,8 +11,12 @@ public class MainLayout : Layout
 {
 	public MainLayout()
 	{
-		Label titleLabel = new(new PixelBounds(896, 256, 128, 128), "Main Layout", LabelStyle.Default with { TextColor = Color.Yellow });
-		TextButton scrollContentButton = new(new PixelBounds(256, 320, 256, 128), () => Game.Self.ActiveLayout = Game.Self.ScrollContentLayout, ButtonStyle.Default, TextButtonStyle.Default, "ScrollContent");
+		LabelStyle labelStyle = new(Color.Yellow, TextAlign.Left, FontSize.H16);
+		ButtonStyle buttonStyle = new(Color.Black, Color.White, Color.Gray(0.5f), 2);
+		TextButtonStyle textButtonStyle = new(Color.White, TextAlign.Middle, FontSize.H24);
+
+		Label titleLabel = new(new PixelBounds(896, 256, 128, 128), "Main Layout", labelStyle);
+		TextButton scrollContentButton = new(new PixelBounds(256, 320, 256, 128), () => Game.Self.ActiveLayout = Game.Self.ScrollContentLayout, buttonStyle, textButtonStyle, "ScrollContent");
 
 		NestingContext.Add(titleLabel);
 		NestingContext.Add(scrollContentButton);